backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.2
psutil>=5.9.6
chromadb>=0.4.14
ollama>=0.1.3
duckduckgo-search>=3.9.10
html2text>=2024.2.26
python-dotenv>=1.0.0
reportlab>=4.0.7
langchain>=0.1.2
langchain-community>=0.0.14
python-multipart>=0.0.6
aiofiles>=23.2.1
